Time: 48 ms (78.41%), Space: 22.4 MB (6.67%) - LeetHub

+++ 1283-find-the-smallest-divisor-given-a-threshold/1283-find-the-smallest-divisor-given-a-threshold.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     int smallestDivisor(vector<int>& nums, int threshold) {
-        long long int n=nums.size();
+        int n=nums.size();
         int maxi=INT_MIN;
         for(auto x:nums)
             maxi=max(x,maxi);
         long long int start=1,end=maxi;
         while(start<end){
-            long long int mid=start+(end-start)/2;
+            int mid=start+(end-start)/2;
             long long int sum=0;
             for(auto x:nums)
                 sum+=ceil((float)x/(float)mid);
